requirements.txt
aniso8601==7.0.0
asgiref==3.2.7
autopep8==1.5.2
Django==3.0.6
django-cors-headers==3.3.0
django-db==0.0.7
django-graphql-jwt==0.3.1
graphene==2.1.8
graphene-django==2.10.1
graphql-core==2.3.2
graphql-relay==2.0.1
promise==2.3
pycodestyle==2.6.0
PyJWT==1.7.1
PyMySQL==0.9.3
pytz==2020.1
Rx==1.6.1
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.3.1
